backend/requirements.txt
fastapi
uvicorn
numpy
sentence-transformers
redis
python-dotenv
requests
pytest
httpx
cyborgdb
